RAIN-2783 Changed datatype of limit and offset

+++ egov-hrms/src/main/java/org/egov/hrms/repository/EmployeeQueryBuilder.java
@@ -83,7 +83,7 @@ public class EmployeeQueryBuilder {
 			pagination = pagination.replace("$offset", "0");
 		
 		if(null != criteria.getLimit()){
-			Long limit = criteria.getLimit() + criteria.getOffset();
+			Integer limit = criteria.getLimit() + criteria.getOffset();
 			pagination = pagination.replace("$limit", limit.toString());
 		}
 		else
